refactor: tobe 패키지 제거 및 모듈 이동 개선
