+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/ControllerScannerTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 import com.interface21.context.stereotype.Controller;
+import com.interface21.webmvc.servlet.mvc.ControllerScanner;
 import java.util.Map;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
